requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2
django-filter==24.2
djangorestframework==3.15.2
idna==3.7
Markdown==3.6
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
